Added "components"
A component, inside a code, is a particular variable we are solving
the problem for. For example, currently we solve the system to get
the value of u, so the only component is u.

Moreover, added the possibility to specify, for the boundary
conditions, which component the condition refers to