Improved components
Changed the components.h file. Now, to retrieve an iterable object
with all the components, there is a function named all_components()
that generates a std::set with all the components. This substitutes
the old array AllComponents[];

The main advantage of using a set intead of an array is that we can
rely on the size() method to retrieve the total number of
components (otherwise, we were forced to divide the size of the
array by the size of on element of the enum).

Finally, a function to get the explicit name of the components has
been added.