Fixed the set_current_solution function
Finally, the code is completely compliant with the fact that we
have three components now. The last piece that was broken was the
set_current_solution function, which is the method that set the
value of the solution starting from a known function.

Now, this method relies on the function set_component, that
works separately for each component.

Moreover, I add a test for this methods, which is called
SetComponentMethod.

Finally, for writing the test I have been forced to extend the
methods to check the error between the current solution and
another known function. Therefore, this commit also introduces the
methods estimate_{l2|linfty|h1}_error for the displacements and the
methods estimate_{l2|linfty}_error_on_trace for the traces.