Renamed functions and variables
An inheritance of the old code was some strange names for the
variables and for the functions.

The suffix "local" that before was used to specify that something
was defined on the cells and not on the trace, now has been
replaced by the suffix "cell". On the other side, for the other
objects tha suffix "trace" has been added. Now, the suffix "local"
means "for just one cell, and not for the overall problem".

Also, the function for assembling the matrix for one cell have been
renamed: now the function that assemble the part of the matrix
generated by test functions defined on the cell and by sample
functions defined on the trace will be called "generate_ct_matrix",
where c is for "cell" and t" is for trace.