Removed the code that set E=0 after the charge neutrality
The equation for the charge neutrality only computes the value of V.
So, the decision is if we must try to approximate E as the projection
of the gradient of V or set E to zero. The decision is not so trivial
because the solution of the local charge neutrality has a lot of
discontinuities and, therefore, E looks like a sum of Dirac deltas.

In any case, for the time being, I removed the code that set E to 0
and I keep the projection