Unfied the python interface among different dimensions
Before, there were three python libraries (called pyddhdg1d,
pyddhdg2d and pyddhdg3d), each one of them with its own objects
that were not compatible with the others.

Now there is just one library, called pyddhdg and the objects
that do not dipend on the dimension are in common.