Added the possibility to set phi_n and phi_p
Added two new components: phi_n and phi_p. Now it is possible
to specify, in the project_component method, phi_n and phi_p as
components. The values of n and p will be changed accordingly so
that the new n (coupled with the current V) generate the requested
value for phi_n in every point (and the same for p and phi_p).