Improved Dirichlet boundary conditions
This commit introduces a new way for computing Dirichlet boundary
conditions. Instead of trying to obtain the L2 projection of the
conditions in the global system, now we compute the projection
cell by cell.

Hopefully, this will be more stable: the boundary condition will
still be correct even if the global linear system has some numerical
errors and the global matrix will be invariant for scaling