Solver is templates on the problem class
Before this commit, a Solver had two template parameters. The
dimension and the Permittivity used. The same was true for a Problem;
the solver required as input a problem with the same parameters.

Now a problem has 4 template parameters: the dimension, the
permittivity but also the electron mobility and the hole mobility.
In this way, it is possibile to change the mobilities without too
much effort. Indeed, now the mobilities classes completely hide
their implementation and just return the result of apply the mu
operator on a vector. In this way, we do not need to store an
array of 2d tensor when mu is just a scalar.

Finally, the solver now has 2 parameters: the first one is the
dimension and the second one is not the Permittivity anymore, but
is the class used for the Problem. In this way, the solver gets
the types of the permittivity and of the mobilities from the
problem.