Added method assemble_system and get_residual for python interface
This commit add some new functions for the python interface. The
first one is the method "assemble_system" that allows to assemble
the system for the trace without resolving the linear system.

Another method is "get_residual" to get the residual of the system
that has just been assembled. Of course, if this method is called
after the "assemble_system" method it will return the residual of
the system that has been just assembled. Otherwise, it will
return the residual of the last iteration of the "run" method.
If none of these methods has been called, it will return an array
with uninitialized values.

Finally, there is a method called "get_dirichlet_boundary_dofs"
that returns (for each component) a couple of vectors. The first
one contains the indices of the dofs that are constrained by the
Dirichlet boundary conditions. The other one contains the values
that these dofs must assume (in the trace update)