Added the possibility to use a different flux for Drift diffusion
Added a flag (that can be specified inside the NPSolverParameters
constructor) that specifies what kind of flux must be used.

"use_cell" means that the values of n on the faces of the cells
when computing the flux will be taken from the solution on the
cells.
"use_trace", instead, means to use the values from the trace