Added "linearize_phi" mode
When this mode is enabled, the Poisson equation is linearized
respect to phi_n and phi_p instead that on n and p.

Be careful that the equations for n and p do not change the values
for phi_n and phi_p (during a single computation step) and
therefore this flags split decouples the system; the poisson
equation does not depend on n and p (because uses phi_n and phi_p
computed at the previous steps) and then n and p are solved using
the new value of V. This is a little bit incoherent; therefore,
this mode should be used only when n and p are disabled.