Removed stabilized tau also from n and p    WARNING: Maybe bugged!
For some reason, one test fails when introducing this commit. The
error was small (from 1e-6 to 3e-6) and the convergence was in any
case granted (if we increase the size of the cells). But, because
I can not find any reason why this should happen, maybe this
commit deserves a double check