Introduced CellFaceTauNPSolverParameter
Added the class CellFaceTauNPSolverParameter, that allows to solve
a problem specifying two different values of tau for each face of
the triangulation (depending on the current cell).

This commit also slightly modify the CMakeFiles. Indeed, now the
GoogleTest package is embedded inside this project and therefore
it is not an external dependency anymore.

Finally, this also splits the dealii-python-bindings package,
creating a static library with the symbols.