Fixing again tests CMakeLists.txt file
I am trying again of obtain something decent for that
CMakeLists.txt file; i.e. a cmake script that is able to compile
the googletest library if it is missing and it is able to use the
system one if it is present.
Right now, it seems that the include(GoogleTest) command is useful
only when the code must be compiled and it does not work when the
code is installed by the user. Moreover, the name of the library
changes from "gtest" to "GTest:GTest" when installed from cmake.