Now get_solution return a rescaled function
If get_solution method is called specifing a component, the output
is a solution rescaled accordingly with the rescaling factor set
in the adimensionalizer.

Unfortunately, this is not True if the commands is called without
a component (and, therefore, if it must return a function with all
the values of the system). This should be fixed, but it is
complicated because it involves multiply each component of the
solution by a different constant.